backend/requirements.txt
# Core
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# GeoDjango
GDAL==3.4.1

# Authentication
django-allauth==0.58.2
djangorestframework-simplejwt==5.3.0

# CORS
django-cors-headers==4.3.0

# Environment variables
python-dotenv==1.0.0

# API Documentation
drf-yasg==1.21.7

# Performance and Caching
django-redis==5.4.0

# Task Queue
celery==5.3.5

# Search
django-elasticsearch-dsl==7.3.0

# Image Processing
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Code Quality
flake8==6.1.0
black==23.11.0

# Production
gunicorn==21.2.0

# Monitoring
sentry-sdk==1.34.0

# MongoDB
mongoengine==0.27.0
django-rest-framework-mongoengine==3.4.1